mssql: splitting to match the convention
