+++ azurerm/internal/services/mssql/validate/elastic_pool_id.go
