+++ azurerm/internal/services/mssql/validate/sql_virtual_machine_login_username_test.go
@@ -2,7 +2,7 @@ package validate
 
 import "testing"
 
-func TestMsSqlVMLoginUserName(t *testing.T) {
+func TestSqlVirtualMachineLoginUserName(t *testing.T) {
 	testCases := []struct {
 		input       string
 		shouldError bool
@@ -17,7 +17,7 @@ func TestMsSqlVMLoginUserName(t *testing.T) {
 	}
 
 	for _, test := range testCases {
-		_, es := MsSqlVMLoginUserName(test.input, "name")
+		_, es := SqlVirtualMachineLoginUserName(test.input, "name")
 
 		if test.shouldError && len(es) == 0 {
 			t.Fatalf("Expected validating name %q to fail", test.input)
